requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
google-cloud-storage==2.9.0
google-cloud-bigquery==3.11.4
functions-framework==3.4.0
